app/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
PyYAML==6.0
redis==4.5.4
requests==2.26.0
python-multipart==0.0.6